Django_Filter/requirements.txt
Django==1.11
django-filter==1.1.0
django-registration-redux==2.0
MySQL-python==1.2.5
pytz==2017.3
